requirements.txt
# ADAPT Engineering Cockpit — Python dependencies
# Compatible with Streamlit Community Cloud (Python 3.11)
# Install: pip install -r requirements.txt

# ── Web framework ──────────────────────────────────────────────────────────────
streamlit>=1.36.0,<2

# ── Data layer ─────────────────────────────────────────────────────────────────
pandas>=2.2.0,<3
pyarrow>=16.0.0,<20
python-dateutil>=2.9.0,<3

# ── Charts ─────────────────────────────────────────────────────────────────────
plotly>=5.22.0,<6

# ── HTTP / Jira API ────────────────────────────────────────────────────────────
requests>=2.31.0,<3

# ── AI advisory (optional — disabled unless COCKPIT_AI_ENABLED=true) ──────────
anthropic>=0.28.0,<1

# ── Environment / config ───────────────────────────────────────────────────────
python-dotenv>=1.0.0,<2

# ── Performance (recommended on Streamlit Cloud) ───────────────────────────────
watchdog>=4.0.0,<5
